+++ dart/utils/XmlHelpers.cpp
@@ -34,6 +34,7 @@
 
 #include <iostream>
 #include <vector>
+
 #include <boost/algorithm/string.hpp>
 #include <boost/lexical_cast.hpp>